backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic[email]==2.6.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
statsmodels==0.14.1
apscheduler==3.10.4
alembic==1.13.1
prometheus-client==0.20.0